Revert ImageResize module
