+++ submodules/FakePasscodeUI/BUILD
@@ -21,7 +21,7 @@ swift_library(
         "//submodules/PasscodeUI:PasscodeUI",
         "//submodules/PasscodeInputFieldNode:PasscodeInputFieldNode",
         "//submodules/AvatarNode:AvatarNode",
-        "//submodules/ImageResize:ImageResize",
+        "//submodules/LocalMediaResources:LocalMediaResources",
     ],
     visibility = [
         "//visibility:public",
