--- submodules/ImageResize/BUILD
@@ -1,17 +1,0 @@
-load("@build_bazel_rules_swift//swift:swift.bzl", "swift_library")
-
-swift_library(
-    name = "ImageResize",
-    module_name = "ImageResize",
-    srcs = glob([
-        "Sources/**/*.swift",
-    ]),
-    copts = [
-        "-warnings-as-errors",
-    ],
-    deps = [
-    ],
-    visibility = [
-        "//visibility:public",
-    ],
-)
